+++ features/step_definitions/chef.rb
@@ -1,0 +1,3 @@
+Given(/I generate a cookbook named '(.+)'/) do |cbname|
+  run_simple "chef generate cookbook #{cbname}"
+end