+++ spec/support/fixtures/code_generator_2/recipes/cookbook.rb
@@ -1,1 +1,3 @@
-# [PH]
+template 'README.md' do
+  source 'README_md.erb'
+end