+++ History.md
@@ -1,5 +1,10 @@
 # Changelog for chef-gen-template
 
+## 0.4.0
+
+* added a GitInit snippet that runs 'git init .'
+* added a simple Aruba tests that verifies that 'chef generate cookbook foo' works
+
 ## 0.3.0
 
 * renamed from ChefDK::Template::Plugin to ChefGen::Flavors at Chef's request so as to not pollute the ChefDK namespace